+++ cmd/dep_radar/main.go
@@ -1,43 +1,94 @@
 package main
 
 import (
+	"context"
 	"encoding/json"
-	"flag"
 	"fmt"
-	"io/ioutil"
+	"log"
+	"net/http"
 	"os"
+	"strings"
 
-	"github.com/stamm/dep_radar/src"
-	"github.com/stamm/dep_radar/src/helpers"
+	"github.com/stamm/dep_radar"
+	"github.com/stamm/dep_radar/app"
+	"github.com/stamm/dep_radar/deps"
+	"github.com/stamm/dep_radar/html"
+	"github.com/stamm/dep_radar/html/templates"
+	"github.com/stamm/dep_radar/providers"
+	"github.com/stamm/dep_radar/providers/github"
+)
+
+const (
+	port = "8081"
 )
 
 func main() {
-	recommendedFile := flag.String("recommended_file", "", "path to file with recommended versions of libraries")
-	githubToken := flag.String("github_token", "", "token for github: maybe empty, but limit for requested applies")
-	githubOrg := flag.String("github_org", "", "name for github organization")
-	listen := flag.String("listen", ":8081", "on which ip and port http server will be listen to")
-	flag.Parse()
-	if *listen == "" {
-		fmt.Println("parameter listen is empty")
-		os.Exit(1)
+	http.HandleFunc("/", handlerPage)
+	http.HandleFunc("/api/", handlerAPI)
+	http.HandleFunc("/favicon.ico", http.NotFound)
+	log.Printf("Started: http://localhost:%s/\n", port)
+	http.ListenAndServe(":"+port, nil)
+}
+
+func handlerPage(w http.ResponseWriter, r *http.Request) {
+	raw, err := templates.Asset("html/templates/main.html")
+	if err != nil {
+		http.Error(w, fmt.Sprintf("err: %s", err), http.StatusBadRequest)
+		return
 	}
-	if *githubOrg == "" {
-		fmt.Println("parameter github_org is empty")
-		os.Exit(1)
+	w.Write(raw)
+}
+
+func handlerAPI(w http.ResponseWriter, r *http.Request) {
+	w.Header().Add("Access-Control-Allow-Origin", "*")
+	githubProv := github.New(github.NewHTTPWrapper(os.Getenv("GITHUB_TOKEN"), 10))
+	provDetector := providers.NewDetector().AddProvider(githubProv)
+	depDetector := deps.DefaultDetector()
+
+	apps := make(chan dep_radar.IApp, 10)
+	orgName := r.URL.Query().Get("github_org")
+	if orgName == "" {
+		http.Error(w, "empty name", http.StatusBadRequest)
+		return
 	}
 
-	var recom src.MapRecommended
-	if *recommendedFile != "" {
-		raw, err := ioutil.ReadFile(*recommendedFile)
+	ctx := context.Background()
+	var (
+		pkgs []dep_radar.Pkg
+		err  error
+	)
+	if strings.Contains(orgName, "/") {
+		pkgs = append(pkgs, dep_radar.Pkg(github.Prefix+"/"+orgName))
+	} else {
+		pkgs, err = githubProv.GetAllOrgRepos(context.Background(), orgName)
 		if err != nil {
-			fmt.Printf("error on read file %s: %s\n", *recommendedFile, err.Error())
-			os.Exit(1)
-		}
-		err = json.Unmarshal(raw, &recom)
-		if err != nil {
-			fmt.Printf("error on unmarshal json: %s\n", err.Error())
-			os.Exit(1)
+			log.Printf("error for getting repos: %s\n", err)
+			http.Error(w, fmt.Sprintf("error while getting repos: %s\n", err), http.StatusBadRequest)
+			return
 		}
 	}
-	helpers.GithubOrg(*githubToken, *githubOrg, *listen, recom)
+	go func() {
+		defer close(apps)
+		for _, pkg := range pkgs {
+			apiApp, err := app.New(ctx, pkg, "master", provDetector, depDetector)
+			if err != nil {
+				log.Printf("cant create app %s, got err: %s\n", pkg, err)
+			}
+			apps <- apiApp
+		}
+	}()
+
+	var recom dep_radar.MapRecommended
+	// 	err = json.Unmarshal(raw, &recom)
+	// 	if err != nil {
+	// 		fmt.Printf("error on unmarshal json: %s\n", err.Error())
+	// 		os.Exit(1)
+	// 	}
+	data := html.Prepare(ctx, apps, provDetector, recom)
+	result, err := json.Marshal(data)
+	if err != nil {
+		http.Error(w, fmt.Sprintf("error %s", err), http.StatusInternalServerError)
+		return
+	}
+	w.Write(result)
 }